refactor(api): split handlers_health_checks.go (2385 → 9 files) by concern
Split the worst red-flag Go file (#761) into nine same-package files by
protocol/concern, matching niac/go's #534-#539 worst-first refactor pattern.
Function bodies are untouched; only relocated.

  handlers_health_checks.go              237  entry handler + shared types + getTestStatus
  handlers_health_checks_settings.go     414  settings GET/PUT + per-endpoint builders
  handlers_health_checks_settings_types.go 193  TestsSettingsResponse + 14 endpoint types
  health_checks_ping.go                  142  PingStats + runExtendedPing + ping driver
  health_checks_port.go                  175  TCP/UDP/DNS-53 port reachability
  health_checks_http.go                  576  HTTP test paths + timing + body/redirect
  health_checks_http_security.go         222  per-header validation (HSTS, CSP, etc.)
  health_checks_tls.go                   142  CertInfo + cert-expiry check + TLS version
  health_checks_rtsp.go                  119  RTSP OPTIONS reachability (#778)
  health_checks_dicom.go                 234  DICOM A-ASSOCIATE-RQ reachability (#777)

All files now under the 600-line file-size warn budget. Drive-by cleanups
forced by --new-from-rev lint of the relocated lines:

- Replace `map[string]string{"status":"success",...}` settings ack with a
  typed `statusResponse{}` using the existing `statusSuccess` constant.
- Rename the shadowing fallback variable in `runStandardHTTPPath` from
  `httpTimings` (shadowed the type) to `fallbackTimings`.

The DICOM A-ASSOCIATE PDU framing keeps its existing int→byte shift+mask
(protocol-bounded 16 KB max PDU). The lint warning is suppressed at the
config level for that file only, with a documented justification — this
behaviour pre-dates the refactor and is exactly what the DICOM spec
requires.

Build, vet, and package tests (SKIP_NETWORK_TESTS=1) all pass.
